requirements.txt
antlr4-python3-runtime==4.8
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cycler==0.10.0
datasets==1.8.0
dill==0.3.4
filelock==3.0.12
fsspec==2021.6.1
huggingface-hub==0.0.10
hydra-core==1.1.0
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
multiprocess==0.70.12.2
numpy==1.21.0
omegaconf==2.1.0
packaging==20.9
pandas==1.2.5
Pillow==8.2.0
pyarrow==3.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
six==1.16.0
threadpoolctl==2.1.0
tqdm==4.49.0
typing-extensions==3.10.0.0
urllib3==1.26.5
xxhash==2.0.2
